refactor: make `exportAccount`, `listRequests` and `getRequest` optional
The async transaction flow shouldn't require snaps to implement the
`listRequests` and `getRequest` methods, so we are making them optional.

Also, the `exportAccount` may not be supported by all snaps, so it
doesn't make sense to force it's implementation.

We decided to keep the CRUD account management methods mandatory, since
the are (or will be) directly called by MetaMask.