feat: remove `buildHandlersChain`
This function lacks many features we expect from a middleware manager,
like allowing the inspection of the response. Also, it uses an exception
to check if the chain should stop or continue, which can be considered a
bad design.

However, we add a new `isKeyringRpcMethod` function that allow us to
check if a method is a keyring RPC method.