feat: remove `KeyringSnapControllerClient` to fix dependency problems (#241)
The `KeyringSnapControllerClient` class depends on the `SnapController`
which adds a lot of dependencies to the `keyring-api`, causing circular
dependency problems in other packages that depend on the `keyring-api`.

The `KeyringSnapControllerClient` is currently used only by the
`eth-snap-keyring` package, which is internal to MetaMask, so this
change shouldn't break any other dependent code.